made changes to the structure of how the audio file is shared within components. [Can be redux in the future?], working play button and intended behavior happening. Next target: implementing behavior for next/prev buttons
